+++ src/libmicroemu/internal/executor/instr/unary_load_instr_imm.h
@@ -43,7 +43,7 @@ public:
 
     TRY_ASSIGN(data, ExecResult, TOp::Read(ictx, address));
 
-    if (arg_t.Get() == static_cast<u8>(RegisterId::kPc)) {
+    if (arg_t.Get() == RegisterId::kPc) {
       // When the given address was unaligend the behaviour is
       // unpredtictable
       if ((address & 0x3u) == 0u) {
